refactor: move static variable USER_TOKEN_CLAIM_PREFIX into common class
