+++ zeebe/auth/src/main/java/io/camunda/zeebe/auth/api/JwtAuthorizationBuilder.java
@@ -19,7 +19,6 @@ public interface JwtAuthorizationBuilder<T extends JwtAuthorizationBuilder<T, A,
   public static String DEFAULT_ISSUER = "zeebe-gateway";
   public static String DEFAULT_AUDIENCE = "zeebe-broker";
   public static String DEFAULT_SUBJECT = "zeebe-client";
-  public static final String USER_TOKEN_CLAIM_PREFIX = "user_token_";
 
   /**
    * Sets the token subject.
