+++ zeebe/auth/src/main/java/io/camunda/zeebe/auth/impl/Authorization.java
@@ -11,6 +11,7 @@ public class Authorization {
 
   public static final String AUTHORIZED_TENANTS = "authorized_tenants";
   public static final String AUTHORIZED_USER_KEY = "authorized_user_key";
+  public static final String USER_TOKEN_CLAIM_PREFIX = "user_token_";
 
   public static JwtAuthorizationEncoder jwtEncoder() {
     return new JwtAuthorizationEncoder();
